organize transpose tests
